Implement code refactoring for chenges model. Move changes type into separate model
